move sum to mocha sample
